backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
openai==1.58.1
dbt-semantic-layer==0.7.10
python-dotenv==1.0.1

